Добавлен абстрактный класс и наследование от него
