+++ src/view/trip-point-container-view.js
@@ -1,25 +1,11 @@
-import {createElement} from '../render.js';
+import AbstractView from './abstract-view.js';
 
 const createTripPointListTemplate = () => (
   '<ul class="trip-events__list"></ul>'
 );
 
-export default class TripPointContainer {
-  #element = null;
-
-  get element() {
-    if (!this.#element) {
-      this.#element = createElement(this.template);
-    }
-
-    return this.#element;
-  }
-
+export default class TripPointContainer extends AbstractView {
   get template() {
     return createTripPointListTemplate();
-  }
-
-  removeElement() {
-    this.#element = null;
   }
 }
